# fluence/openapi.yaml
openapi: 3.0.0
info:
  title: Fluence
  version: 0.1.0
servers:
  - url: '/v1'
...
components:
  schemas:
    Account:
      type: object
      properties:
        stark_key:
          type: string
        address:
          type: string
          nullable: true
      required: [stark_key]
    Blueprint:
      type: object
      properties:
        permanet_id:
          type: string
          nullable: true
        minter:
          $ref: '#/components/schemas/Account'
        expire_at:
          type: string
          nullable: true
      required: [minter]
    Collection:
      type: object
      properties:
        address:
          type: string
        fungible:
          type: boolean
        blueprint:
          $ref: '#/components/schemas/Blueprint'
        name:
          type: string
        symbol:
          type: string
        decimals:
          type: integer
        image:
          type: string
          nullable: true
      required: [address, fungible, name, symbol, decimals]
    Metadata:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        image:
          type: string
      required: [name, description, image]
    Token:
      allOf:
        - $ref: '#/components/schemas/Metadata'
        - type: object
          properties:
            contract:
              $ref: '#/components/schemas/Collection'
            token_id:
              type: string
          required: [contract, token_id]
...
    TransferRequest:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        amount_or_token_id:
          type: string
        contract:
          type: string
        nonce:
          type: string
      required: [from, to, amount_or_token_id, contract, nonce]
...
    Tx:
      type: object
      properties:
        transaction_hash:
          type: string
      required: [transaction_hash]
paths:
...
  /clients:
    post:
      operationId: register_client
      summary: Register client ethereum address
      description: Register a stark key and a matching ethereum address.
      tags: [client]
      parameters:
        - name: signature
          in: query
          required: true
          description: |
            `sign([address, nonce])`
          schema:
            type: array
            items:
              type: string
          explode: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                stark_key:
                  type: string
                address:
                  type: string
                nonce:
                  type: string
              required: [stark_key, address, nonce]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tx'
  /clients/{address}:
    get:
      operationId: get_client
      summary: Get client stark key
      description: Find client stark key by ethereum address.
      tags: [client]
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  stark_key:
                    type: string
                required: [stark_key]
        '404':
          description: The address is not registered yet.
  /blueprints:
    post:
      operationId: create_blueprint
      summary: Setup a permanent id
      description: Setup a permanent id and the minter.
      tags: [collection]
      parameters:
        - name: signature
          in: query
          required: true
          description: |
            `sign([permanent_id])`
          schema:
            type: array
            items:
              type: string
          explode: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                permanent_id:
                  type: string
                minter:
                  type: string
              required: [permanent_id, minter]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  permanent_id:
                    type: string
                  minter:
                    $ref: '#/components/schemas/Account'
                  expire_at:
                    type: string
                required: [permanent_id, minter, expire_at]
  /collections:
    get:
      operationId: find_collections
      summary: Find collections
      description: Find collections by owner.
      tags: [collection]
      parameters:
        - name: owner
          in: query
          description: Ethereum address
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Collection'
                  total:
                    type: integer
                required: [data, total]

    post:
      operationId: register_collection
      summary: Register collection
      description: |
        Register a collection and construct a forward registration request with signature.
      tags: [collection]
      parameters:
        - name: signature
          in: query
          description: |
            `sign([address, name, symbol, base_uri, image])`
          schema:
            type: array
            items:
              type: string
          explode: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
              allOf:
                - type: object
                  properties:
                    address:
                      type: string
                    name:
                      type: string
                    symbol:
                      type: string
                    base_uri:
                      type: string
                    image:
                      type: string
                  required: [address, name, symbol, base_uri, image]
                - oneOf:
                    - type: object
                      properties:
                        blueprint:
                          type: string
                      required: [blueprint]
                    - type: object
                      properties:
                        minter:
                          type: string
                      required: [minter]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  req:
                    type: object
                    properties:
                      from:
                        type: string
                      to:
                        type: string
                      value:
                        type: string
                      gas:
                        type: string
                      batch:
                        type: string
                      nonce:
                        type: string
                      data:
                        type: string
                    required: [from, to, value, gas, batch, nonce, data]
                  signature:
                    type: string
                required: [req, signature]
  /_metadata/{permanent_id}/{token_id}:
    get:
      operationId: get_metadata_by_permanent_id
      summary: Get token metadata
      description: Get token metadata by (collection) permanent ID and token ID.
      tags: [metadata]
      parameters:
        - name: permanent_id
          in: path
          required: true
          schema:
            type: string
        - name: token_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '404':
          description: The token is not found.
  /collections/{address}/tokens/{token_id}/_metadata:
    parameters:
      - name: address
        in: path
        required: true
        schema:
          type: string
      - name: token_id
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: get_metadata
      summary: Get token metadata
      description: Get token metadata by collection address and token ID.
      tags: [metadata]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '404':
          description: The token is not found.

    put:
      operationId: update_metadata
      summary: Update token metadata
      description: Update token metadata by collection address and token ID.
      tags: [metadata]
      parameters:
        - name: signature
          in: query
          required: true
          description: |
            `sign([address, token_id, nonce])`
          schema:
            type: array
            items:
              type: string
          explode: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: |
            The metadata does not conform to the ERC721 metadata JSON Schema.
        '401':
          description: |
            The request is not correctly signed by the registered minter.
        '404':
          description: |
            The collection is not registered yet or not (registered as) mintable.
  /tokens:
    get:
      operationId: find_tokens
      summary: Find tokens
      description: Find tokens by collection and/or owner.
      tags: [token]
      parameters:
        - name: owner
          in: query
          description: Ethereum address
          schema:
            type: string
        - name: collection
          in: query
          description: Contract address
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Token'
                  total:
                    type: integer
                required: [data, total]
  /balance:
    get:
      operationId: get_balance
      summary: Get balance
      description: Get balance by user (stark key) and contract (address).
      tags: [token]
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: string
        - name: contract
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: string
  /owner:
    get:
      operationId: get_owner
      summary: Get token owner
      description: Get owner by contract (address) and token ID.
      tags: [token]
      parameters:
        - name: token_id
          in: query
          required: true
          schema:
            type: string
        - name: contract
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  owner:
                    type: string
...
  /transfer:
    post:
      operationId: transfer
      summary: Transfer token
      description: |
        Transfer token (or tokens) identified by `contract` and `amount_or_token_id`.
      tags: [token]
      parameters:
        - name: signature
          in: query
          required: true
          description: |
            `sign([to, amount_or_token_id, contract, nonce])`
          schema:
            type: array
            items:
              type: string
          explode: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tx'
...